requirements.txt
customtkinter
Pillow
pillow-avif-plugin
fastapi
python-multipart
uvicorn
